refactor: split svg theme auto modules
Separate the --svg-theme-auto parsing and selection helpers from the terminal appearance probe so the binary wiring reads more cleanly.

Keep behavior unchanged by moving the /dev/tty detection into a dedicated terminal_appearance module and leaving the CLI contract and validation coverage intact.
